check for no dotfiles

+++ install.sh
@@ -1,6 +1,10 @@
 #!/bin/sh
 
-mv ~/.profile ~/.profile.old
-mv ~/.kshrc ~/.kshrc.old
+if [ -f "$HOME/.profile" ]; then
+	mv ~/.profile ~/.profile.old
+fi
+if [ -f "$HOME/.kshrc" ]; then
+	mv ~/.kshrc ~/.kshrc.old
+fi
 ln -s ~/.dotfiles/profile ~/.profile
 ln -s ~/.dotfiles/kshrc ~/.kshrc
